fix broken build.

+++ NaLib/NaString.cpp
@@ -236,7 +236,7 @@ int NaString::CompareNoCase(const wchar_t * lpsz) const
 
 int NaString::Find(const wchar_t * const ch, const int begin /*= 0*/) const
 {
-	int itemlen = wcslen(ch);
+	int itemlen = (int)wcslen(ch);
 	if (m_nLen < itemlen)
 		return -1;
 
@@ -377,7 +377,7 @@ int NaString::ReplaceAll(wchar_t * from, wchar_t * to)
 		temp += Mid(begin, index - begin);
 		temp += to;
 
-		begin = index + wcslen(from);
+		begin = index + (int)wcslen(from);
 	}
 
 	SetBuf(temp.wstr());
@@ -516,7 +516,7 @@ const NaString & NaString::SetBuf(const wchar_t* lpsz)
 	}
 	else
 	{
-		m_nLen = wcslen(lpsz);
+		m_nLen = (int)wcslen(lpsz);
 		m_nBufLen = (int)(sizeof(wchar_t) * (m_nLen + 1));
 
 		AllocBuf(m_nBufLen);
@@ -619,7 +619,7 @@ int NaStrArray::Remove(const int nIndex)
 
 int NaStrArray::GetCount() const
 {
-	return m_Array.size();
+	return (int)m_Array.size();
 }
 
 int NaStrArray::Find(const NaString &str)
